Move mix node detail components to a location that is better named
